safari-600.7-branch/LayoutTests/crypto/subtle/rsa-export-key-expected.txt

Test exporting an RSA key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Importing a JWK key...
PASS crypto.subtle.exportKey(null, key) threw exception TypeError: Unknown key format.
PASS crypto.subtle.exportKey(undefined, key) threw exception TypeError: Unknown key format.
PASS crypto.subtle.exportKey({}, key) threw exception TypeError: Unknown key format.
PASS crypto.subtle.exportKey("", key) threw exception TypeError: Unknown key format.
PASS crypto.subtle.exportKey("foobar", key) threw exception TypeError: Unknown key format.

Exporting the key as JWK...
PASS exportedJWK.kty is 'RSA'
PASS exportedJWK.n is publicKeyJSON.n
PASS exportedJWK.e is publicKeyJSON.e
PASS exportedJWK.alg is 'RS256'
PASS exportedJWK.ext is true
PASS exportedJWK.use is undefined
PASS exportedJWK.key_ops is ['sign', 'verify']
PASS successfullyParsed is true

TEST COMPLETE
